refactor MainActivity
