+++ app/src/main/java/com/metroyar/classes/Path.kt
@@ -1,4 +1,4 @@
-package com.metroyar
+package com.metroyar.classes
 
 import com.metroyar.model.Station
