+++ app/src/main/java/com/metroyar/classes/Result.kt
@@ -1,11 +1,13 @@
-package com.metroyar
+package com.metroyar.classes
 
 import android.content.Context
-import com.metroyar.GlobalObjects.metroGraph
-import com.metroyar.GlobalObjects.tripleOfLinesAndTheirStartAndEndStations
+import com.metroyar.R
+import com.metroyar.utils.GlobalObjects.metroGraph
+import com.metroyar.utils.GlobalObjects.tripleOfLinesAndTheirStartAndEndStations
 import com.metroyar.model.Station
 import com.metroyar.utils.findStationObjectFromItsId
 import com.metroyar.utils.findStationObjectFromItsName
+import com.metroyar.utils.getDirectionFromInterchangeStations
 import com.metroyar.utils.log
 import java.util.PriorityQueue
