+++ app/src/main/java/com/metroyar/screen/AccountScreen.kt
@@ -1,0 +1,13 @@
+package com.metroyar.screen
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+
+@Composable
+fun AccountScreen() {
+    Column(horizontalAlignment = Alignment.CenterHorizontally) {
+        Text("Account")
+    }
+}
