+++ app/src/main/java/com/metroyar/utils/DirectionFromInterchangeNodes.kt
@@ -1,4 +1,4 @@
-package com.metroyar
+package com.metroyar.utils
 
 import com.metroyar.utils.findStationObjectFromItsId
