+++ app/src/main/res/drawable/baseline_near_me_24.xml
@@ -1,5 +1,5 @@
 <vector android:height="24dp" android:tint="#000000"
     android:viewportHeight="24" android:viewportWidth="24"
     android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M13,10l5,0l3,-3l-3,-3l-5,0l0,-2l-2,0l0,2l-7,0l0,6l7,0l0,2l-5,0l-3,3l3,3l5,0l0,4l2,0l0,-4l7,0l0,-6l-7,0z"/>
+    <path android:fillColor="@android:color/white" android:pathData="M21,3L3,10.53v0.98l6.84,2.65L12.48,21h0.98L21,3z"/>
 </vector>
